// app/types/order-types/order-types.ts
export interface OrderData {
  user_id: string;
  user_name: string;
  user_email: string;
  status: string;
  total_amount: number;
  product_id: number;
  product_name: string;
  quantity: number;
  unit_price: number;
  size: string;
  product_category_id: number;
  product_category_name: string;
  design_category_id: number;
  design_category_name: string;
  phone_number: string;
  alt_phone_number: string;
  address: string;
  city: string;
}